MetaLater: Invoke later callbacks queued by earlier ones
If a MetaLater callback queued another MetaLater with a scheduling
later than the one currently being invoked, make it so that the newly
scheduled callback will actually be invoked.

The fact that it doesn't already do this is a regression from
cd7a9680932868d1d2ef5447c77712cef9a443dd.

https://bugzilla.gnome.org/show_bug.cgi?id=755605