smartcardManager: Add way to detect if user logged using token
If a user uses a token at login time and they have there system
configured to lock on smartcard removal we need to make sure they continue
to use the token at unlock time.

As a prerequisite for addressing that problem we need to know up front
if a user logged in with a token at all.

This commit adds the necessary api to detect that case.